Move code to application_operator
